plan_exec/pyproject.toml
[project]
name = "plan-exec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "pydantic-settings>=2.11.0",
    "tavily-python>=0.7.12",
]
